# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Content, enquiry and admin API for Acquire 360 Ventures Ltd
 * OpenAPI spec version: 0.1.0
 */

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @nullable */
  logo?: string | null;
}